# docs/openapi.yaml
openapi: 3.1.0
info:
    title: RSTC API
    version: 1.0.0
    description: REST API for RSTC application.
servers:
    - url: /api
paths:
...
    /users/self/self-password:
        put:
            tags:
                - Users
            summary: Update own password
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - currentPassword
                                - newPassword
                            properties:
                                currentPassword:
                                    type: string
                                newPassword:
                                    type: string
                            examples:
                                change:
                                    value:
                                        currentPassword: OldPass1
                                        newPassword: NewPass1
            responses:
                '200':
                    description: Password updated
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
                            examples:
                                success:
                                    value:
                                        success: true
                '400':
                    $ref: '#/components/responses/Error'
                    description: Invalid input
                '401':
                    $ref: '#/components/responses/Error'
                    description: Current password incorrect
                '404':
                    $ref: '#/components/responses/Error'
                    description: User not found
                '500':
                    $ref: '#/components/responses/Error'
                    description: Server error
    /users:
        get:
            tags:
                - Users
            summary: List all users
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Users list
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/User'
                            examples:
                                list:
                                    value:
                                        - id: 1
                                          username: admin
                                          role: admin
                                          permissions:
                                              - users:view
                                          status: active
                                          last_login: '2026-07-27T00:00:00.000Z'
                                          login_count: 1
                                          created_at: '2026-07-27T00:00:00.000Z'
                '500':
                    $ref: '#/components/responses/Error'
        post:
            tags:
                - Users
            summary: Create user
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserCreate'
                        examples:
                            admin:
                                value:
                                    username: newuser
                                    password: Pass1234
                                    role: user
                                    permissions:
                                        - users:view
            responses:
                '200':
                    description: User created
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
                '400':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
    /users/{id}:
        put:
            tags:
                - Users
            summary: Update user
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserUpdate'
                        examples:
                            update:
                                value:
                                    username: updateduser
                                    role: user
                                    status: active
            responses:
                '200':
                    description: User updated
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
                '400':
                    $ref: '#/components/responses/Error'
                '404':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
        delete:
            tags:
                - Users
            summary: Delete user
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: User deleted
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
                '400':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
    /users/{id}/password:
        put:
            tags:
                - Users
            summary: Update user password
            security:
                - bearerAuth: []
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - password
                            properties:
                                password:
                                    type: string
                                    example: NewPass123
            responses:
                '200':
                    description: Password updated
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
                '400':
                    $ref: '#/components/responses/Error'
                '500':
                    $ref: '#/components/responses/Error'
...
components:
    schemas:
...
        User:
            type: object
            properties:
                id:
                    type: integer
                username:
                    type: string
                role:
                    type: string
                permissions:
                    type: array
                    items:
                        type: string
                status:
                    type: string
                last_login:
                    type: string
                    format: date-time
                login_count:
                    type: integer
                created_at:
                    type: string
                    format: date-time
        UserCreate:
            type: object
            required:
                - username
                - password
            properties:
                username:
                    type: string
                    example: newuser
                password:
                    type: string
                    example: Pass1234
                role:
                    type: string
                    example: user
                permissions:
                    type: array
                    items:
                        type: string
                    example:
                        - users:view
        UserUpdate:
            type: object
            properties:
                username:
                    type: string
                    example: updateduser
                role:
                    type: string
                    example: user
                status:
                    type: string
                    example: active
                permissions:
                    type: array
                    items:
                        type: string
...
        SuccessResponse:
            type: object
            properties:
                success:
                    type: boolean
        ErrorResponse:
            type: object
            properties:
                error:
                    type: string
    responses:
        Error:
            description: Error response
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
                    examples:
                        badRequest:
                            value:
                                error: 'Invalid request'
                        unauthorized:
                            value:
                                error: 'دسترسی غیرمجاز! لطفاً ابتدا وارد شوید.'
                        forbidden:
                            value:
                                error: 'دسترسی غیرمجاز'
                        notFound:
                            value:
                                error: 'یافت نشد'
                        serverError:
                            value:
                                error: 'خطای سرور'